kern: log SEV-SNP guestmode
NB:  The definition SEV_STAT_SNP_ACTIVE is straight from the AMD
documentation.  The use of "active" instead of "enabled" -- as with
the other definitions -- feels inconsistent.  However, "SNP enabled"
in the context of SNP means the host system is able to run VMs using
SNP.  And "SNP active" means, the guest is actually using SNP.  And
not just e.g. SEV-ES.