SEV-ES guest: install final #VC trap handler
Install the actual #VC entry stub.

As out instructions used to ack, mask and unmask interrupts in the
PIC, we will raise #VC while handling interrupts.  Most exceptions
re-enable interrupts as soon as the stack frame is setup etc.  And
will disable them again before unwinding the stack and return from
the exception.  However, in case of #VC we want to have interrupts
disabled during the exception handling.  Otherwise we would get
nested IRQs of the same PSL.

Therefore do not use the TRAP() macro but rewrite the entry stub
without enabling interrupts and branch to the "common" code as soon
as possible.