vmm(4):  For SEV-ES enabled guest raise #VC on MMIO related #NPF
This is accomplished by setting up a mapping in the nested page
table, that has reserved bits set.  The bits of the physical address
portion of the PTE that are affected bit physcial address bit
reduction are reserved.  When these bits are set, a #VC trap with
error code #NPF will be raised.

For non-SEV-ES guests forward the #NPF to vmd(8).